+++ service/grunt/clean.js
@@ -1,5 +1,0 @@
-module.exports = {
-  coverage: {
-    src: ['test/coverage/']
-  }
-};