+++ service/grunt/env.js
@@ -1,5 +1,0 @@
-module.exports = {
-  coverage: {
-    APP_DIR_FOR_CODE_COVERAGE: '../test/coverage/instrument/lib/'
-  }
-};