+++ service/grunt/instrument.js
@@ -1,7 +1,0 @@
-module.exports = {
-  files: 'lib/**/*.js',
-  options: {
-    lazy: true,
-    basePath: 'test/coverage/instrument/'
-  }
-};