+++ service/grunt/makeReport.js
@@ -1,8 +1,0 @@
-module.exports = {
-  src: 'test/coverage/reports/**/*.json',
-  options: {
-    type: 'html',
-    dir: 'test/coverage/reports',
-    print: 'detail'
-  }
-};