+++ service/grunt/node-inspector.js
@@ -1,8 +1,0 @@
-module.exports = {
-  dev: {
-    options: {
-      'web-port': 8081,
-      'debug-port': 5858
-    }
-  }
-};