+++ service/grunt/storeCoverage.js
@@ -1,5 +1,0 @@
-module.exports = {
-  options: {
-    dir: 'test/coverage/reports'
-  }
-};